requirements.txt
discord.py
youtube-dl
PyNaCl
ffmpeg-python